[#3] Refactor: 차트 만드는 const 및 util들 리펙터링 및 비즈니스 로직 hook으로 리펙터링
