+++ src/utils/chartUtils.ts
@@ -1,0 +1,37 @@
+export function createCustomTooltip({
+  opt,
+  timeList,
+  barValueList,
+  areaValueList,
+  idList,
+}: {
+  opt: any;
+  timeList: string[];
+  barValueList: number[];
+  areaValueList: number[];
+  idList: string[];
+}) {
+  const index = opt.dataPointIndex;
+  return `
+						<ul class='arrow-box'>
+							<li class='arrow-box__item'>
+								${timeList[index]}
+							</li>
+							<li class='arrow-box__item'>
+								<div style="background:${opt.w.globals.colors[0]}; width:10px; height:10px; border-radius:10px"></div>
+								<div>bar: </div>
+								<div>${barValueList[index]}</div>
+							</li>
+							<li class='arrow-box__item'>
+								<div style="background:${opt.w.globals.colors[1]}; width:10px; height:10px; border-radius:10px"></div>
+								<div>area: </div>
+								<div>${areaValueList[index]}</div>
+							</li>
+							<li class='arrow-box__item'>
+								
+								<div>지역: </div>
+								<div>${idList[index]}</div>
+							</li>
+						</ul>
+					`;
+}
